implementing storage as library
started implementing storage as its own library. i'm in the middle of
modifying saveSelectedContent.js to use the new library, or at least
save to storage consistently. I don't know how to inject two scripts
in a page and have them place nice. It might just work, without any
forethought (it could be as simple as jump dumping the script in).
I want to inject my new Storage library and have saveSelectedContent
work with that.